Extract test class definition to separate files
