+++ test/support/base.rb
@@ -1,0 +1,3 @@
+class ActiveSupport::TestCase
+  include FactoryGirl::Syntax::Methods
+end
