+++ test/support/controllers.rb
@@ -1,0 +1,3 @@
+class ActionController::TestCase
+  include Devise::Test::ControllerHelpers
+end
