+++ test/support/integration.rb
@@ -1,0 +1,22 @@
+require 'capybara/rails'
+require 'capybara/minitest'
+
+class ActionDispatch::IntegrationTest
+  include Warden::Test::Helpers
+  include Capybara::DSL
+  include Capybara::Minitest::Assertions
+
+  teardown do
+    Capybara.reset_sessions!
+  end
+end
+
+class JsFeatureTest < ActionDispatch::IntegrationTest
+  setup do
+    Capybara.current_driver = Capybara.javascript_driver
+  end
+
+  teardown do
+    Capybara.use_default_driver
+  end
+end
